Remove serialization subproject and use sbt/serialization
