--- protocol-test/src/test/scala/sbt/protocol/SpecsUtil.scala
@@ -1,17 +1,0 @@
-package sbt.protocol.spec
-
-import sbt.protocol.Message
-import scala.pickling.PickleOps, sbt.serialization._, sbt.serialization.json._
-import scala.pickling.static._
-import scala.pickling.Defaults.pickleOps
-
-object SpecsUtil {
-  import sbt.serialization.spec.JUnitUtil._
-
-  // TODO get rid of pickleMessage and parseMessage once pickle/unpickle are not macros
-  def pickleMessage(m: Message): String =
-    SerializedValue(m).toJsonString
-
-  def parseMessage(s: String): Message =
-    SerializedValue.fromJsonString(s).parse[Message].get
-}
