--- serialization/src/main/scala/sbt/serialization/pickler/SerializedValue.scala
@@ -1,6 +1,0 @@
-package sbt.serialization
-package pickler
-
-trait SerializationPicklers {
-  implicit val serializedValuePickler: SPickler[SerializedValue] with Unpickler[SerializedValue] = SerializedValue.pickler
-}
