--- serialization/src/main/scala/sbt/serialization/pickler/package.scala
@@ -1,7 +1,0 @@
-package sbt.serialization
-
-package object pickler {
-  type PrimitivePicklers = scala.pickling.pickler.PrimitivePicklers
-  type PrimitiveArrayPicklers = scala.pickling.pickler.PrimitiveArrayPicklers
-  type RefPicklers = scala.pickling.pickler.RefPicklers
-}
